requirements.txt
Pillow==9.0.0
pre-commit==2.17.0
